mini_app/requirements.txt
PySide6>=6.5
requests>=2.28
pytesseract>=0.3.10
Pillow>=9.0
